# parallelise gen-dir calls