much of the time in `bazel run //:gen_srcs` is IO and so can be parallelised for a significant speedup.  parallelising the calls to `gen-dir` speeds up my benchmark of `-main` from ~7s to 5s

the overall `bazel run //:gen_srcs` runtime drops from 9.7s to 7.2s (although these are quite variable).